requirements.txt
streamlit
pdf2image
pillow
requests
transformers[torch]
git+https://github.com/ndgnuh/vietocr